Fix typo in initial velocity
Everywhere else speaks of 12,000 yards a second; this is the only place
that speaks of one-tenth that velocity. This seems a particularly
unfortunate error as it is part of the response from the authoritative
Cambridge Observatory.

Note that https://gutenberg.org/cache/epub/12901/pg12901.txt agrees with
the current "initial velocity of 1,200 yards a second" at this location,
but the French version at https://gutenberg.org/cache/epub/38674/pg38674.txt
gives it as "vitesse initiale de douze mille yards par seconde" (rather
than "douze cents").